% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilMoistureTemperatureRegimes.R
\name{calc_SMTRs}
\alias{calc_SMTRs}
\title{Calculate soil moisture and soil temperature regimes and underlying
conditions}
\usage{
calc_SMTRs(
  sim_in,
  sim_out = NULL,
  sim_agg = NULL,
  soil_TOC = NULL,
  has_soil_temperature = TRUE,
  opt_SMTR = list(aggregate_at = "conditions", crit_agree_frac = 0.9, use_normal = TRUE,
    SWP_dry = -1.5, SWP_sat = -0.033, impermeability = 0.9),
  simTime1 = NULL,
  simTime2 = NULL,
  verbose = FALSE,
  msg_tag = NULL
)
}
\arguments{
\item{sim_in}{An object of class
\code{\link[rSOILWAT2:swInputData-class]{swInputData}}.
The \pkg{rSOILWAT2} simulation input.}

\item{sim_out}{An object of class
\code{\link[rSOILWAT2:swOutput-class]{swOutput}}.
The \pkg{rSOILWAT2} simulation output. If \code{NULL} then \code{sim_agg}
must be provided instead.}

\item{sim_agg}{A named list. The prepared \pkg{rSOILWAT2} simulation output.
If \code{NULL} then \code{sim_out} must be provided so that the elements
of \code{sim_agg} can be determined internally. The list contains:
\var{"soiltemp.dy.all"}, \var{"soiltemp.yr.all"}, \var{"soiltemp.mo.all"},
\var{"vwcmatric.dy.all"}, \var{"swpmatric.dy.all"}, \var{"prcp.yr"},
\var{"prcp.mo"}, \var{"pet.mo"}, and \var{"temp.mo"}. Note:
if \code{sim_agg} has already been calculated for other reasons, then
passing \code{sim_agg} may be faster than passing \code{sim_out},
which (re-)calculates \code{sim_agg}.}

\item{soil_TOC}{A numeric vector. Total soil organic matter in g C / kg soil
for each soil layer. If \code{NULL}, then internally set to 0.}

\item{has_soil_temperature}{A logical value. Set to \code{TRUE}, if soil
temperature values have been simulated.}

\item{opt_SMTR}{A named list. Parameters for the calculation of the regimes:
\itemize{
  \item \code{aggregate_at}: A character string. Determines the approach
    for regime determination; we recommend the value of "conditions".
    See details.
  \item \code{crit_agree_frac}: A numeric value. The aggregation agreement
    level (e.g., 0.5 = majority; 1 = all); we recommend a value of 0.9.
  \item \code{use_normal}: A logical value. If \code{TRUE}, then
    "normal years" as defined by (Soil Survey Staff 2014: p.29) are
    determined (note: should be at least a time period of 30 years);
    if \code{FALSE}, then all years are used for the calculation of
    regimes. We recommend to determine "normal years".
  \item \code{SWP_dry}: A numeric value. Soil water potential (MPa)
    below which soils are considered dry; we recommend a value of -1.5 MPa.
  \item \code{SWP_sat}: A numeric value. Soil water potential (MPa)
    above which soils are considered saturated; we recommend a value
    of -0.033 MPa.
  \item \code{impermeability}: A numeric value. The value above which
    the code considers a soil layer to be impermeable. We recommend a
    value of 0.9.
}}

\item{simTime1}{A list with named elements. Calculated internally
if \code{NULL}; alternatively, it can be generated by a call to the
function \code{\link[rSW2data]{setup_time_simulation_run}}.}

\item{simTime2}{A list with named elements. Calculated internally
if \code{NULL}; alternatively, it can be generated by a call to the
function \code{\link[rSW2data]{simTiming_ForEachUsedTimeUnit}}.}

\item{verbose}{A logical value. If \code{TRUE} more messages are printed.}

\item{msg_tag}{A character string. Tag that is pre-appended to each verbose
message.}
}
\value{
A list with the following elements: \itemize{
  \item \code{regimes_done}: if successful calculation of soil moisture and
    soil temperature regimes then \code{TRUE} otherwise \code{FALSE}
  \item \code{has_simulated_SoilTemp}: if soil temperature was simulated
    then \code{1} otherwise \code{0}
  \item \code{has_realistic_SoilTemp}: if simulated soil temperature values
    are realistic then \code{1} otherwise \code{0}
  \item \code{has_Ohorizon}: if soil profile may have an O-horizon, then
    \code{TRUE} otherwise \code{FALSE}
  \item \code{Lanh_depth}: a numeric vector of length two
  \item \code{MCS_depth}: a numeric vector of length two
  \item \code{Fifty_depth}: a numeric value
  \item \code{permafrost_yrs}: number of years with permafrost conditions
  \item \code{SMR_normalyears}: years considered "normal"
  \item \code{SMR_normalyears_N}: number of years considered "normal"
  \item \code{cond_annual}: a numeric matrix with annual values of
    underlying conditions used to determine soil moisture and soil
    temperature regimes
  \item \code{STR}: soil temperature regimes
  \item \code{SMR}: soil moisture regimes
}
}
\description{
Calculations are based on SSS (2014, 2015) and explained in detail in the
\code{vignette(
  topic = "SoilMoistureRegimes_SoilTemperatureRegimes",
  package = "rSW2funs"
)}.
}
\section{Details}{
 The argument \code{aggregate_at} determines at which level
  aggregations to regimes are carried out: \itemize{
  \item \code{data}: Determine conditions and regimes based on aggregated
    mean soil moisture/temperature values.
  \item \code{conditions} Determine conditions based on time-series values
    of soil moisture/temperature values; determine regimes based on
    aggregated mean conditions.
  \item \code{regime}: Determine conditions and regimes based on
    time-series values of soil moisture/temperature values.
}
}

\examples{
sw_in <- rSOILWAT2::sw_exampleData
sw_out <- rSOILWAT2::sw_exec(inputData = sw_in)
SMTR <- rSW2funs::calc_SMTRs(sim_in = sw_in, sim_out = sw_out)

}
\references{
Soil Survey Staff (2014). Keys to soil taxonomy,
  12th ed. USDA Natural Resources Conservation Service, Washington, DC.

Soil Survey Staff (2015). Illustrated guide to soil
  taxonomy. USDA Natural Resources Conservation Service, National Soil Survey
  Center, Lincoln, Nebraska.
}
