% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{default_parameters_GISSM_bigsagebrush}
\alias{default_parameters_GISSM_bigsagebrush}
\title{Default parameter values of GISSM for big sagebrush}
\usage{
default_parameters_GISSM_bigsagebrush()
}
\description{
Default parameter values of GISSM for big sagebrush
}
