% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GISSM_kill_seedling}
\alias{GISSM_kill_seedling}
\title{Set seedling as dead for a given day in a given year (\var{\sQuote{ss1s}})}
\usage{
GISSM_kill_seedling(ss1s, ry_year_day, ry_useyrs, y, doy)
}
\arguments{
\item{ss1s}{Logical vector. Elements represent calendar days included in
all calendar year of \code{ry_useyrs}.}

\item{ry_year_day}{Numerical vector. Elements represent calendar days
and values represent the calendar year of each day.}

\item{ry_useyrs}{Numerical vector. The sequence of calendar year.}

\item{y}{Numerical value. The index of the currently examined calendar year.}

\item{doy}{Numerical value. The index of the currently examined calendar day.}
}
\description{
Set seedling as dead for a given day in a given year (\var{\sQuote{ss1s}})
}
\section{Note}{
 The \pkg{Rcpp} version of the function is about 270x faster
 for vectors of length 365 and 12,000x faster for vectors of length 11,000
 than the R version. The \pkg{Rcpp} version also reduced the memory
 footprint by a factor of >> 3080.

 Previous name \code{setFALSE_SeedlingSurvival_1stSeason}.
}

\section{C code}{
 \code{ss1s} is a pointer to the data and the original
 vector will get altered; one would need for a deep copy:
 \code{LogicalVector out = clone(ss1s)}
}

\examples{
 # The \pkg{Rcpp} function is equivalent to the following R version
   kill_seedling_R <- function(ss1s, ry_year_day, ry_useyrs, y,
     doy) {
     ss1s[ry_year_day == ry_useyrs[y]][doy] <- FALSE
     ss1s
   }

}
\references{
Schlaepfer, D.R., Lauenroth, W.K. & Bradford, J.B. (2014).
 Modeling regeneration responses of big sagebrush (Artemisia tridentata)
 to abiotic conditions. Ecol Model, 286, 66-77.
}
