% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{parameters_GISSM_bigsagebrush}
\alias{parameters_GISSM_bigsagebrush}
\title{Parameter values of GISSM for big sagebrush}
\usage{
parameters_GISSM_bigsagebrush(...)
}
\arguments{
\item{...}{Named GISSM parameters and their values.}
}
\value{
A list of default and adjusted parameter values for GISSM.
}
\description{
Parameter values of GISSM for big sagebrush
}
\examples{
# Default values
x <- parameters_GISSM_bigsagebrush()

# Fix \var{Doy_SeedDispersalStart0} to 325 and use default values otherwise
x <- parameters_GISSM_bigsagebrush(Doy_SeedDispersalStart0 = 325)

}
\references{
Schlaepfer, D.R., Lauenroth, W.K. & Bradford, J.B. (2014).
  Modeling regeneration responses of big sagebrush (Artemisia tridentata)
  to abiotic conditions. Ecol Model, 286, 66-77.
}
