% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilMoistureTemperatureRegimes.R
\name{SMR_logic}
\alias{SMR_logic}
\title{\var{NRCS} soil moisture regimes}
\usage{
SMR_logic(
  ACS_COND1,
  ACS_COND2,
  ACS_COND3,
  MCS_COND0,
  MCS_COND1,
  MCS_COND2,
  MCS_COND2_1,
  MCS_COND2_2,
  MCS_COND2_3,
  MCS_COND3,
  MCS_COND3_1,
  MCS_COND4,
  MCS_COND5,
  MCS_COND6,
  MCS_COND6_1,
  MCS_COND7,
  MCS_COND8,
  MCS_COND9,
  MCS_COND10,
  has_permafrost
)
}
\description{
Soil moisture regimes are determined based on
Soil Survey Staff 2014 (pages 28-31, Key to Soil Taxonomy).
}
\references{
Soil Survey Staff (2014). Keys to soil taxonomy,
  12th ed. USDA Natural Resources Conservation Service, Washington, DC.
}
\keyword{internal}
