% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{calc_GISSM}
\alias{calc_GISSM}
\title{The germination and individual seedling survival model \var{GISSM}}
\usage{
calc_GISSM(
  x,
  soillayer_depths_cm,
  params = parameters_GISSM_bigsagebrush(),
  site_latitude = NULL,
  has_soil_temperature = NULL,
  simTime1 = NULL,
  simTime2 = NULL,
  plot = FALSE,
  path = NULL,
  filename_tag = "GISSM"
)
}
\description{
GISSM represents the frequency of years when big sagebrush seeds germinate
and seedlings survive in undisturbed natural vegetation
(Schlaepfer et al. 2014).
}
