% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{calc_GISSM}
\alias{calc_GISSM}
\title{The germination and individual seedling survival model \var{GISSM}}
\usage{
calc_GISSM(
  x,
  soillayer_depths_cm,
  params = parameters_GISSM_bigsagebrush(),
  site_latitude = NULL,
  has_soil_temperature = NULL,
  years = NULL,
  simTime1 = NULL,
  simTime2 = NULL,
  debug_output = 0L,
  path = NULL,
  filename_tag = "GISSM"
)
}
\arguments{
\item{x}{A named list or an object of
\pkg{rSOILWAT2} class \code{\linkS4class{swOutput}} with daily output.
If \code{x} is a named list, then it must contain appropriate content for
\var{SWP_MPa}, \var{Snowpack_SWE_mm}, \var{air_Tmin_C}, \var{air_Tmax_C},
\var{air_Tmean_C}, \var{shallowsoil_Tmin_C}, \var{shallowsoil_Tmean_C},
\var{shallowsoil_Tmax_C}. See examples.}

\item{soillayer_depths_cm}{A numeric vector. The lower bounds of simulated
soil layers.}

\item{params}{A named list. See \code{\link{parameters_GISSM_bigsagebrush}}.}

\item{site_latitude}{A numeric value.
Required if \code{simTime2} is \code{NULL}.}

\item{has_soil_temperature}{A logical value or \code{NULL}. Optional
information whether or not object \code{x} contains (good)
soil temperature values.}

\item{years}{A numeric vector or \code{NULL}. The sequence of simulated
calendar years.
extracted from \code{x}
if \pkg{rSOILWAT2} class \code{\linkS4class{swOutput}};
otherwise, required if \code{simTime1} is \code{NULL}.}

\item{simTime1}{A named list or \code{NULL}.
See \code{\link[rSW2data]{setup_time_simulation_run}}.
If \code{NULL}, then derived from \code{years};
otherwise, must be consistent with \code{years}.}

\item{simTime2}{A named list or \code{NULL}.
See \code{\link[rSW2data]{simTiming_ForEachUsedTimeUnit}}.
If \code{NULL}, then derived from \code{simTime1};
otherwise, must be consistent with \code{simTime1}.}

\item{debug_output}{An integer value. Level of additional outputs.
If \code{0}, then standard variables are returned.
If \code{1} or \code{2}, then additional elements are included in the
returned item; see \var{Value}.
If \code{2}, then additional output for debugging purposes is
written to a \var{csv} spreadsheet and a \var{pdf} figure is created.}

\item{path}{A character string. The path to the directory where additional
output should be written to disk.}

\item{filename_tag}{A character string. File name without extension used
for writing additional output.}
}
\value{
A named list with one element: \describe{
    \item{outcome}{A \var{data.frame} tabulating success/failure
    for \var{Germination_Emergence} and for \var{SeedlingSurvival_1stSeason}
    for each regeneration year.}
  }

  If \code{debug_output} is \code{2}, then seven additional elements
  are provided for debugging purposes: \describe{
    \item{successes_days}{Number of days per year with positive outcomes.}
    \item{success_mostfrequent_doy}{
      Seasonal timing of most frequent positive outcomes.
    }
    \item{time_to_germinate_days}{Time to germinate.}
    \item{nogermination_days}{Days without germination.}
    \item{nogermination_periods_yrs}{Consecutive years without germination.}
    \item{noseedlings_periods_yrs}{
      Consecutive years without seedling recruitment.
    }
    \item{mortality_causes}{Causes of seedling mortality.}
  }
}
\description{
GISSM represents the frequency of years when big sagebrush seeds germinate
and seedlings survive in undisturbed natural vegetation
(Schlaepfer et al. 2014).
}
\section{Notes}{

  Time information is checked for consistency only minimally;
  unexpected output may be produced if
  \code{x}, \code{years}, \code{simTime1}, or \code{simeTim2} contain
  inconsistent time content.
}

\examples{
sw_in <- rSOILWAT2::sw_exampleData
res <- rSOILWAT2::sw_exec(inputData = sw_in)

# Example 1: use rSOILWAT2 output directly to run GISSM
GISSM_r1 <- calc_GISSM(
  x = res,
  soillayer_depths_cm = rSOILWAT2::swSoils_Layers(sw_in)[, 1],
  site_latitude = rSOILWAT2::swSite_IntrinsicSiteParams(sw_in)[["Latitude"]],
  has_soil_temperature =
    rSOILWAT2::swSite_SoilTemperatureFlag(sw_in) &&
    !rSOILWAT2::has_soilTemp_failed()
)


# Example 2: use list of daily values to run GISSM
#   populate daily values from rSOILWAT2 output as here or from other model
dt <- "Day"
tmp_swp <- slot(slot(res, "SWPMATRIC"), dt)
tmp_snow <- slot(slot(res, "SNOWPACK"), dt)
tmp_airtemp <- slot(slot(res, "TEMP"), dt)
tmp_soiltemp <- slot(slot(res, "SOILTEMP"), dt)

has_sl_minmeanmax <- grepl(
  "Lyr_1_avg_C",
  colnames(tmp_soiltemp),
  fixed = TRUE
)
cns_sl <- if (any(has_sl_minmeanmax)) {
  # rSOILWAT2 since v5.3.0
  paste0("Lyr_1_", c("min", "avg", "max"), "_C")
} else {
  # rSOILWAT2 before v5.3.0
  # Daily mean soil temperature in the absence of daily min/max
  rep("Lyr_1", 3)
}

GISSM_r2 <- calc_GISSM(
  x = list(
    SWP_MPa = -1 / 10 * tmp_swp[, -(1:2), drop = FALSE],
    Snowpack_SWE_mm = 10 * tmp_snow[, "snowpackWaterEquivalent_cm"],
    air_Tmin_C = tmp_airtemp[, "min_C"],
    air_Tmean_C = tmp_airtemp[, "avg_C"],
    air_Tmax_C = tmp_airtemp[, "max_C"],
    shallowsoil_Tmin_C = tmp_soiltemp[, cns_sl[1]],
    shallowsoil_Tmean_C = tmp_soiltemp[, cns_sl[2]],
    shallowsoil_Tmax_C = tmp_soiltemp[, cns_sl[3]]
  ),
  soillayer_depths_cm = rSOILWAT2::swSoils_Layers(sw_in)[, 1],
  site_latitude = rSOILWAT2::swSite_IntrinsicSiteParams(sw_in)[["Latitude"]],
  years =
    rSOILWAT2::swYears_StartYear(sw_in):rSOILWAT2::swYears_EndYear(sw_in)
)

all.equal(GISSM_r1, GISSM_r2)

# Calculate the frequency of years when big sagebrush seeds germinate
# and seedlings survive in undisturbed natural vegetation
# (Schlaepfer et al. 2014)
colMeans(GISSM_r1[["outcome"]][, -1])

}
\references{
Schlaepfer, D.R., Lauenroth, W.K. & Bradford, J.B. (2014).
  Modeling regeneration responses of big sagebrush (Artemisia tridentata)
  to abiotic conditions. Ecol Model, 286, 66-77.
}
