% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2funs-package.R
\docType{package}
\name{rSW2funs}
\alias{rSW2funs}
\alias{rSW2funs-package}
\title{\pkg{rSW2funs}: Collection of functions that calculate new variables
using rSOILWAT2 simulation output.}
\description{
Collection of functions that calculate new variables
    using rSOILWAT2 simulation output.
}
\section{LICENSE}{

   Copyright (C) \Sexpr{format(Sys.Date(), "\%Y")} by
   \Sexpr{packageDescription("rSW2funs")[["Maintainer"]]}

   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, version 3 of the License.
}

\section{DISCLAIMER}{

   This program is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
   GNU General Public License for more details.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrylandEcology/rSW2funs}
  \item Report bugs at \url{https://github.com/DrylandEcology/rSW2funs/issues}
}

}
\author{
\strong{Maintainer}: Daniel Schlaepfer \email{danielrschlaepfer@gmail.com} (\href{https://orcid.org/0000-0001-9973-2065}{ORCID})

}
